Moves login and register controllers and models into separate folders
