+++ src/controllers/auth-controllers.ts
@@ -1,25 +1,13 @@
 import { RequestHandler } from "express"
-import { registerUser, logInUser } from "../models/auth-models"
-import { generateToken } from "../middleware/authentication"
 
 
-export const postRegistration: RequestHandler = async (req, res, next) => {
+export const testAuth: RequestHandler = async (_req, res, next) => {
 
   try {
-    const user = await registerUser(req.body)
-    res.status(201).send({ user })
-  } catch (err) {
-    next(err)
-  }
-}
-
-
-export const postLogin: RequestHandler = async (req, res, next) => {
-
-  try {
-    const user = await logInUser(req.body)
-    const token = generateToken(user)
-    res.status(200).send({ token })
+    res.status(200).send({
+      message: "OK",
+      details: "Authentication successful"
+    })
   } catch (err) {
     next(err)
   }
