+++ src/models/register-models.ts
@@ -1,5 +1,5 @@
 import { db } from "../db"
-import { Credentials, LoggedInUser, SecureUser, User } from "../types/user-types"
+import { SecureUser, User } from "../types/user-types"
 import { checkEmailConflict } from "../utils/db-query-utils"
 
 
@@ -37,45 +37,6 @@ export const registerUser = async ({ username, password, email, first_name, surn
     `,
     [username, password, email, first_name, surname, unit_preference]
   )
-
-  return result.rows[0]
-}
-
-
-export const logInUser = async ({ username, password }: Credentials): Promise<LoggedInUser> => {
-
-  const result = await db.query(`
-    SELECT 
-      user_id, 
-      username 
-    FROM users 
-    WHERE username = $1;
-    `,
-    [username])
-
-  if (!result.rowCount) {
-    return Promise.reject({
-      status: 404,
-      message: "Not Found",
-      details: "Username could not be found"
-    })
-  }
-
-  const dbPassword = await db.query(`
-    SELECT password
-    FROM users
-    WHERE username = $1
-    AND password = $2;
-    `,
-    [username, password])
-
-  if (!dbPassword.rowCount) {
-    return Promise.reject({
-      status: 401,
-      message: "Unauthorized",
-      details: "Incorrect password"
-    })
-  }
 
   return result.rows[0]
 }
